+++ CHANGELOG.md
@@ -20,7 +20,9 @@ Int64 value.
 ZcashRustBackendWelding:
 - `func getReceivedMemoAsUTF8(dbData:idNote:networkType:) -> String?`
  Use `getReceivedMemo(dbData:idNote:networkType)` instead
- 
+- `func getSentMemoAsUTF8(dbData:idNote:networkType:) -> String?`
+ Use `getSentMemo(dbData:idNote:networkType)` instead
+
 ### [#209] Support Initializer Aliases
 
 Added `ZcashSynchronizerAlias` enum which is used to identify an instance of the `SDKSynchronizer`. All the paths